+++ includes/class.llms.ajax.php
@@ -79,7 +79,11 @@ class LLMS_AJAX {
 		// script will only register once
 		wp_register_script( 'llms',  plugins_url( '/assets/js/llms' . LLMS_Frontend_Assets::$min . '.js', LLMS_PLUGIN_FILE ), array( 'jquery' ), '', true );
 		wp_localize_script( 'llms', 'wp_ajax_data', $this->get_ajax_data() );
-		wp_enqueue_script( 'llms' );
+
+		// ensure this doesn't load on the wp-login.php screen
+		if ( false === stripos( $_SERVER['SCRIPT_NAME'], strrchr( wp_login_url(), '/' ) ) ) {
+			wp_enqueue_script( 'llms' );
+		}
 
 	}